socialnet/requirements.txt
Django==5.0.2
djangorestframework-simplejwt==5.3.1
pyjwt==2.9.0 
mysqlclient==2.0.3 
django-environ==0.8.1